backend/pyproject.toml
[tool.poetry]
name = "laxmi-pm-backend"
version = "0.1.0"
description = ""
authors = ["Abhishek Mahajan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
python-multipart = "^0.0.20"
pydantic = "^2.10.3"
uvicorn = "^0.34.0"
numpy = "^2.2.0"
pandas = "^2.2.3"
azure-search = "^1.0.0b2"
tiktoken = "^0.8.0"
openai = "^1.58.0"
azure-ai-formrecognizer = "^3.3.3"
azure-search-documents = "^11.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
